requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
peewee==3.17.0
pydantic==2.5.2
requests==2.31.0
typing-extensions==4.9.0
urllib3==1.26.13
telebot==0.0.5
loguru~=0.7.2
python-telegram-bot-calendar==1.0.5
